requirements.txt
argh==0.26.2
asn1crypto==0.24.0
bcrypt==3.1.7
cffi==1.12.3
cryptography==2.7
dateutils==0.6.6
Django==2.2.1
django-background-tasks==1.2.0
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==0.0.8
django-compat==1.0.15
django-rest-framework-braces==0.3.4
django-sslserver==0.20
djangorestframework==3.10.3
djutils
future==0.17.1
numpy==1.16.4
paramiko==2.6.0
pathtools==0.1.2
pexpect==4.7.0
ptyprocess==0.6.0
py-dateutil==2.2
pycparser==2.19
pydicom==1.3.0
PyNaCl==1.3.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
scpclient==0.7
six==1.12.0
sqlparse==0.3.0
watchdog==0.9.0
pexpect
